+++ packages/design-system/src/components/navigation/SidebarMenuItem.vue
@@ -2,9 +2,12 @@
 import { computed, inject, onBeforeMount, toValue } from 'vue'
 import type { Component } from 'vue'
 import { cn } from '@/lib/utils'
+import {
+  APP_LAYOUT_MENU_INJECTION_KEY,
+  isPinnedSettingsSingleId,
+} from '../layout/appLayoutMenuContext'
 import { SIDEBAR_MENU_INJECTION_KEY } from './sidebarMenuContext'
 import {
-  sidebarMenuChevronClass,
   sidebarMenuIconClass,
   sidebarMenuIconStateClass,
   sidebarMenuLabelClass,
@@ -24,8 +27,30 @@ if (!injectedMenu) {
   throw new Error('SidebarMenuItem must be used inside SidebarMenu')
 }
 
+const layoutMenu = inject(APP_LAYOUT_MENU_INJECTION_KEY, null)
 const menu = injectedMenu
 const active = computed(() => menu.isActive(props.id))
+
+const isPinnedSettingsSingle = computed(() =>
+  layoutMenu
+    ? layoutMenu.settingsMenu.value
+      && isPinnedSettingsSingleId(props.id, layoutMenu.settingsMenuId, menu.parentGroupId)
+    : false,
+)
+
+const renderInSettingsFooter = computed(() => {
+  if (!isPinnedSettingsSingle.value || !layoutMenu) {
+    return false
+  }
+
+  return layoutMenu.settingsSingleTarget.value instanceof HTMLElement
+})
+
+const renderInMain = computed(() => !isPinnedSettingsSingle.value)
+
+const settingsTeleportTarget = computed(
+  () => layoutMenu?.settingsSingleTarget.value ?? document.body,
+)
 
 const classes = computed(() =>
   cn(sidebarMenuTriggerClass(), sidebarMenuStateClass(active.value)),
@@ -39,6 +64,10 @@ const iconClasses = computed(() =>
 )
 
 onBeforeMount(() => {
+  if (isPinnedSettingsSingle.value && layoutMenu) {
+    layoutMenu.registerSettingsSingle()
+  }
+
   const isTopLevel = !menu.parentGroupId
   menu.registerMenuItem(props.id, isTopLevel)
 
@@ -49,16 +78,22 @@ onBeforeMount(() => {
 </script>
 
 <template>
-  <button
-    type="button"
-    :class="classes"
-    :title="menu.collapsed.value ? label : undefined"
-    :aria-current="active ? 'page' : undefined"
-    @click="menu.setActive(id)"
+  <Teleport
+    :disabled="!renderInSettingsFooter"
+    :to="settingsTeleportTarget"
   >
-    <span :class="iconClasses">
-      <component :is="icon" v-if="icon" :size="16" class="shrink-0" />
-    </span>
-    <span :class="sidebarMenuLabelClass()">{{ label }}</span>
-  </button>
+    <button
+      v-show="renderInMain || renderInSettingsFooter"
+      type="button"
+      :class="classes"
+      :title="menu.collapsed.value ? label : undefined"
+      :aria-current="active ? 'page' : undefined"
+      @click="menu.setActive(id)"
+    >
+      <span :class="iconClasses">
+        <component :is="icon" v-if="icon" :size="16" class="shrink-0" />
+      </span>
+      <span :class="sidebarMenuLabelClass()">{{ label }}</span>
+    </button>
+  </Teleport>
 </template>